Refactor revalidation logic by introducing `getPathsByEntity` utility, consolidating tag and path handling, and streamlining imports.
